+++ src/main.js
@@ -7,12 +7,11 @@ import TopRatedComponent from "./components/topRated.js";
 import MostCommentedComponent from "./components/mostCommented.js";
 import StatisticsComponent from "./components/statistics.js";
 import PopupComponent from "./components/popup.js";
-import {FILM_CARDS_PER_ROW, FILM_CARDS_BY_BUTTON, RenderPosition, ESC_KEY} from "./const.js";
+import {EXTRA_FILM_CARDS, FILM_CARDS_PER_ROW, FILM_CARDS_BY_BUTTON, RenderPosition, ESC_KEY} from "./const.js";
 import {generatedFilms} from "./mock/generateFilmCards";
 import {generateMenu} from "./mock/generateMenu";
 import {render} from "./utils.js";
 
-const firstCard = generatedFilms[0];
 const menu = generateMenu(generatedFilms);
 
 let showingFilmsCount = FILM_CARDS_PER_ROW;
@@ -28,7 +27,7 @@ const showMoreBtn = filmComponent.getElement().querySelector(`.films-list__show-
 
 render(header, new ProfileComponent().getElement(), RenderPosition.BEFOREEND);
 render(main, new MenuComponent(menu).getElement(), RenderPosition.BEFOREEND);
-render(main, new FilmComponent().getElement(), RenderPosition.BEFOREEND);
+render(main, filmComponent.getElement(), RenderPosition.BEFOREEND);
 render(footerStatistics, new StatisticsComponent().getElement(), RenderPosition.BEFOREEND);
 
 const renderFilm = (container, film, position) => {
@@ -36,7 +35,6 @@ const renderFilm = (container, film, position) => {
   const popupElement = popupComponent.getElement();
   const popupMainElements = [`.film-card__poster`, `.film-card__title`, `.film-card__comments`];
   const popupCloseBtn = popupElement.querySelector(`.film-details__close-btn`);
-  console.log(popupElement)
 
   const renderPopup = () => {
     body.appendChild(popupElement);
@@ -77,7 +75,6 @@ const renderFilmCollections = (container, films) => {
   films.slice(0, showingFilmsCount)
   .forEach((film) => renderFilm(filmsListContainer, film, RenderPosition.BEFOREEND));
 
-
   showMoreBtn.addEventListener(`click`, () => {
     const prevFilmsCount = showingFilmsCount;
     showingFilmsCount = showingFilmsCount + FILM_CARDS_BY_BUTTON;
@@ -90,23 +87,27 @@ const renderFilmCollections = (container, films) => {
   });
 };
 
-renderFilmCollections(new FilmComponent(), generatedFilms);
+renderFilmCollections(filmComponent, generatedFilms);
 
 
 // Отрисовка доп карточек (topRated и mostCommented)
 
-// const mostCommentedFilms = generatedFilms.slice().sort((a, b) => a.comments.length >= b.comments.length ? -1 : 1);
-// const topRatedFilms = generatedFilms.slice().sort((a, b) => a.rating > b.rating ? -1 : 1);
+
+const topRatedComponent = new TopRatedComponent();
+const mostCommentedComponent = new MostCommentedComponent();
+
+const mostCommentedFilms = generatedFilms.slice().sort((a, b) => a.comments.length >= b.comments.length ? -1 : 1);
+const topRatedFilms = generatedFilms.slice().sort((a, b) => a.rating > b.rating ? -1 : 1);
+render(filmComponent.getElement(), topRatedComponent.getElement(), RenderPosition.BEFOREEND);
+render(filmComponent.getElement(), mostCommentedComponent.getElement(), RenderPosition.BEFOREEND);
 
 
-// const topRatedFilmsList = films.querySelector(`.films-list--extra`);
-// const topRatedFilmsContainer = topRatedFilmsList.querySelector(`.films-list__container`);
+const topRatedFilmsContainer = topRatedComponent.getElement().querySelector(`.films-list__container`);
 
-// topRatedFilms.slice(0, EXTRA_FILM_CARDS)
-// .forEach((film) => render(topRatedFilmsContainer, createFilmCardTemplate(film), `beforeend`));
+topRatedFilms.slice(0, EXTRA_FILM_CARDS)
+ .forEach((film) => renderFilm(topRatedFilmsContainer, film, RenderPosition.BEFOREEND));
 
-// const mostCommentedFilmsList = films.querySelector(`.films-list--extra:last-child`);
-// const mostCommentedFilmsContainer = mostCommentedFilmsList.querySelector(`.films-list__container`);
+const mostCommentedFilmsContainer = mostCommentedComponent.getElement().querySelector(`.films-list__container`);
 
-// mostCommentedFilms.slice(0, EXTRA_FILM_CARDS)
-// .forEach((film) => render(mostCommentedFilmsContainer, createFilmCardTemplate(film), `beforeend`));
+mostCommentedFilms.slice(0, EXTRA_FILM_CARDS)
+.forEach((film) => renderFilm(mostCommentedFilmsContainer, film, RenderPosition.BEFOREEND));
